+++ website/src/examples/fetch/FetchExample.tsx
@@ -37,7 +37,7 @@ function FetchExample() {
   )
 
   const currentUrl = useObservable(url$, '')
-  const response = useObservable(response$)
+  const response = useObservable(response$, null)
 
   return (
     <div>